+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Special thanks to: @rejas, @sdetweil
 - Added test for remoteFile option in compliments module
 - Added hourlyWeather functionality to Weather.gov weather provider
 - Removed weatherEndpoint definition from weathergov.js (not used)
+- Added css class names "today" and "tomorrow" for default calendar
 
 ### Removed